+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 ### Fixed
+- The header/metadata has a more reasonable representation for emptiness
+  when input data is empty.
+
 ### Security
 
 ## [0.14.0] - 2021-05-12